% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally.R
\name{add_count_se}
\alias{add_count_se}
\title{tally/count standard interface.}
\usage{
add_count_se(x, groupingVars = NULL, wt = NULL, sort = FALSE)
}
\arguments{
\item{x}{data.frame to tally/count}

\item{groupingVars}{character vector of column names to group by.}

\item{wt}{character optional column name containing row-weights (passed to count/tally)}

\item{sort}{logical if TRUE sort result in descending order}
}
\value{
.data with added column n, containing counts.
}
\description{
Add a new column named "n" with (optionally per-group) sums/counts.
}
\details{
Note: \code{dplyr::count}, \code{dplyr::add_count},
\code{dplyr::tally}, and \code{dplyr::add_tally} are not \code{S3}
methods, so it may not be practical to re-dispatch \code{seplyr} calls
to these \code{dplyr} implementations.
}
\examples{

suppressPackageStartupMessages(library("dplyr"))

datasets::iris \%>\% count_se(wt = "Sepal.Width", groupingVars= c('Species'))

}
\seealso{
\code{\link[dplyr]{add_count}}
}
