% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_group_indices.R
\name{add_group_indices}
\alias{add_group_indices}
\title{Group a data frame and add per-group indices as a column.}
\usage{
add_group_indices(.data, groupingVars, indexColumn)
}
\arguments{
\item{.data}{data.frame}

\item{groupingVars}{character vector of column names to group by.}

\item{indexColumn}{character name of column to add indices to.}
}
\value{
.data with group identifying column added.
}
\description{
Group a data frame and add per-group indices as a column.
}
\examples{

suppressPackageStartupMessages(library("dplyr"))

add_group_indices(datasets::mtcars, c("cyl", "gear"), 'groupID')

}
