% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedMapBuilder.R
\name{:=}
\alias{:=}
\alias{named_map_builder}
\title{Named map builder.}
\usage{
":="(names, values)
}
\arguments{
\item{names}{names to set.}

\item{values}{values to assign names to (and return).}
}
\value{
values with names set.
}
\description{
Set names of right-argument to be left-argument, and return right argument.
}
\examples{


c('a' := '4', 'b' := '5')
# equivalent to: c(a = '4', b = '5')

c('a', 'b') := c(1, 2)
# equivalent to: c(a = 1, b = 2)

# the important example
name <- 'a'
name := 5
# equivalent to: c("a" = 5)

}
