% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_se.R
\name{deselect}
\alias{deselect}
\title{deselect standard interface.}
\usage{
deselect(.data, colNames)
}
\arguments{
\item{.data}{data.frame}

\item{colNames}{character vector of columns to remove}
}
\value{
.data without deselected columns
}
\description{
deselect columns.  To keep columns please see \code{\link{select_se}}.
}
\examples{

datasets::mtcars \%>\%
   deselect(c("cyl", "gear")) \%>\%
   head()
# essentially dplyr::select( datasets::mtcars, -cyl, -gear)

}
\seealso{
\code{\link{select_se}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{select_at}}
}
