% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_se.R
\name{distinct_se}
\alias{distinct_se}
\title{Standard interface for distinct.}
\usage{
distinct_se(.data, groupingVars, .keep_all = FALSE)
}
\arguments{
\item{.data}{data.frame}

\item{groupingVars}{character vector of column names to group by.}

\item{.keep_all}{logical, passed to dplyr::distinct.}
}
\value{
.data passed through distinct with groupingVars args.
}
\description{
Group a data frame and add per-group indices as a column.
}
\examples{

suppressPackageStartupMessages(library("dplyr"))

datasets::mtcars \%>\% distinct_se(c("cyl", "gear"))

}
\seealso{
\code{\link[dplyr]{distinct}}
}
