% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_nse.R
\name{filter_nse}
\alias{filter_nse}
\title{Filter non-standard interface (deprecated).}
\usage{
filter_nse(.data, ..., filter_nse_env = parent.frame())
}
\arguments{
\item{.data}{data.frame}

\item{...}{stringified expressions to filter by.}

\item{filter_nse_env}{environment to work in.}
}
\value{
.data filtered by columns named in filterTerms
}
\description{
Filter a data frame by the filter terms in \code{...} (deprecated, please use \code{\link{filter_se}}).
}
\seealso{
\code{\link{filter_se}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{filter_at}}
}
