% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_indices_se.R
\name{group_indices_se}
\alias{group_indices_se}
\title{group_indices standard interface.}
\usage{
group_indices_se(.data, groupingVars, add = FALSE)
}
\arguments{
\item{.data}{data.frame}

\item{groupingVars}{character vector of column names to group by.}

\item{add}{logical, passed to group_by}
}
\value{
per-row group index assignments
}
\description{
Group a data frame by the groupingVars and add group labels.
}
\examples{

group_indices_se(datasets::mtcars, c("cyl", "gear"))

}
\seealso{
\code{\link[dplyr]{group_indices}}
}
