% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmtPlanner.R
\name{if_else_device}
\alias{if_else_device}
\title{Simulate a per-row \code{if(){}else{}} block.}
\usage{
if_else_device(testexpr, thenexprs, elseexprs = NULL)
}
\arguments{
\item{testexpr}{character containing the test expression.}

\item{thenexprs}{named character then assignments (altering columns, not creating).}

\item{elseexprs}{named charager else assignments (altering columns, not creating).}
}
\description{
This device uses expression-ifelse to simulate the
more powerful block-ifelse. Note: \code{ifebtest_*}
is a reserved column name for this procedure.
}
\examples{

 # Example: clear one of a or b in any row where both are set.
 d <- data.frame(a = c(0, 0, 1, 1),
                 b = c(0, 1, 0, 1),
                 edited = FALSE)

 program <- if_else_device('(a+b)>1',
                           if_else_device('runif(4) >= 0.5',
                                          c('a' := '0',
                                            edited = TRUE),
                                          c('b' := '0',
                                            edited = TRUE)))
 print(program)

 plan <- partition_mutate_se(program)
 print(plan)

 res <- d
 for(si in plan) {
   res <- mutate_se(res, si)
 }
 res <- res  \%.>\%
   select_se(., grepdf('^ifebtest_.*', ., invert=TRUE))
 print(res)

}
