% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{lambda}
\alias{lambda}
\alias{λ}
\title{Build an anonymous function.}
\usage{
lambda(params, env = parent.frame())

λ(params, body, env = parent.frame())
}
\arguments{
\item{params}{formal parameters of function, unbound names.}

\item{env}{environment to work in}

\item{body}{subsituted body of function to map arguments into.}
}
\value{
user defined function.
}
\description{
Build an anonymous function.
}
\examples{

# example: square numbers
sapply(1:4, lambda(x)(x^2))


# example: square numbers
sapply(1:4, λ(x, x^2))

}
