% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmtPlanner.R
\name{mutate_qt}
\alias{mutate_qt}
\title{Capture the expressions of a mutate-style command.}
\usage{
mutate_qt(...)
}
\arguments{
\item{...}{mutate expressions with := used for assignment.}
}
\value{
ordered list of mutate_se assignment blocks
}
\description{
Capture the expressions of a mutate-style command.
}
\examples{

plan <- mutate_qt(a1 := 1, b1 := a1, a2 := 2, b2 := a1 + a2)
data.frame(x=1) \%.>\% mutate_se(., plan)

}
