% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmtPlanner.R
\name{mutate_seb}
\alias{mutate_seb}
\title{Run a sequence of quoted mutate blocks.}
\usage{
mutate_seb(d, blocks, env = parent.frame())
}
\arguments{
\item{d}{data.frame to work on}

\item{blocks}{list of sequence named char-array of mutate blocks}

\item{env}{environment to work in.}
}
\value{
d with blocks applied in order
}
\description{
Run a sequence of quoted mutate blocks.
}
\examples{

plan <- partition_mutate_qt(a1 := 1, b1 := a1, a2 := 2, b2 := a1 + a2)
print(plan)
d <- data.frame(x = 1) \%.>\% mutate_seb(., plan)
print(d)

}
