% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novelName.R
\name{novelName}
\alias{novelName}
\title{Generate a name with a prefix disjoint from a set of names}
\usage{
novelName(prefix, names)
}
\arguments{
\item{prefix}{character, desired prefix}

\item{names}{character list of names to avoid}
}
\value{
new name disjoint from set of names
}
\description{
Generate a name with a prefix disjoint from a set of names
}
\examples{

# basic op
novelName('b', c('a', 'b', 'c'))

# complex application (converting logistic
# links to probabilities).
d <- data.frame(
  exampleId = c(1, 1, 2, 2),
  resultLabel = c('a', 'b' , 'a', 'b'),
  linkValue = c(-5, 2, -2, -1),
  stringsAsFactors = FALSE)
totColName <- novelName('t', colnames(d))
d ->.;
  mutate_se(., c(totColName := "exp(linkValue)")) ->.;
  group_by_se(., "exampleId") ->.;
  mutate_se(., c("probability" :=
                paste0(totColName, '/sum(', totColName, ')'))) ->.;
  deselect(., totColName)

}
