% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmtPlanner.R
\name{partition_mutate_se}
\alias{partition_mutate_se}
\title{Partition a sequence of mutate commands into longest ordered no create/use blocks.}
\usage{
partition_mutate_se(exprs)
}
\arguments{
\item{exprs}{source of mutate expressions as an assignment list}
}
\value{
ordered list of mutate_se assignment blocks
}
\description{
The idea is: we assume the sequence of expressions is in a valid order
(all items available before use).  This function partitions the expressions
into ordered longest "no new value used blocks" by greedily scanning forward
and skipping any expressions that either use a value created in the current block
or require a value not yet produced.
}
\examples{

partition_mutate_se(c("a1" := "1", "b1" := "a1", "a2" := "2", "b2" := "a1 + a2"))

}
