% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmtPlanner.R
\name{quote_mutate}
\alias{quote_mutate}
\title{Capture the expressions of a mutate-style command.}
\usage{
quote_mutate(...)
}
\arguments{
\item{...}{mutate expressions with := or = used for assignment.}
}
\value{
ordered list of mutate_se assignment blocks
}
\description{
Capture the expressions of a mutate-style command.
}
\examples{

assignments <- quote_mutate(a1 := 1, b1 = a1, a2 := 2, b2 := 7*(a1 + a2))
data.frame(x=1) \%.>\% mutate_se(., assignments)

}
