% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_se.R
\name{rename_mp}
\alias{rename_mp}
\title{rename map interface.}
\usage{
rename_mp(.data, mapping)
}
\arguments{
\item{.data}{data.frame}

\item{mapping}{named character vector of columns to rename (original names on the left, new names on the right; this reverse of dplyr::rename()).}
}
\value{
.data with renamed columns
}
\description{
rename columns (much different syntax than \code{\link[dplyr]{rename}}).
}
\examples{

mapping <- c("cyl" = "cylinders", "gear" = "gears")
datasets::mtcars \%>\%
   rename_mp(mapping) \%>\%
   head()
# # sames as:
# datasets::mtcars \%>\%
#    rename(cylinders = cyl, gears = gear) \%>\%
#    head()

}
\seealso{
\code{\link{rename_se}}, \code{\link[dplyr]{rename}},  \code{\link[dplyr]{rename_at}}
}
