% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_se.R
\name{select_se}
\alias{select_se}
\title{select standard interface.}
\usage{
select_se(.data, colNames)
}
\arguments{
\item{.data}{data.frame}

\item{colNames}{character vector of columns to keep}
}
\value{
.data with only selected columns
}
\description{
select columns.
}
\examples{

datasets::mtcars \%>\% select_se(c("cyl", "gear")) \%>\% head()

}
