% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{seplyr}
\alias{seplyr}
\alias{seplyr-package}
\title{\code{seplyr}: Standard Evaluation Interfaces for Common \code{dplyr} Verbs}
\description{
Supplies replacement interfaces for dplyr that take
variable names as simple strings.  Also supplies some useful "packaged pipelines"
including \code{group_summarize} (also eliminates need for \code{dplyr::add_count} and \code{dplyr::add_tally}), \code{add_group_summaries}, and \code{add_group_indices}.
}
