% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_se.R
\name{summarize_se}
\alias{summarize_se}
\alias{summarise_se}
\title{summarize standard interface.}
\usage{
summarize_se(.data, summarizeTerms)

summarise_se(.data, summarizeTerms)
}
\arguments{
\item{.data}{data.frame}

\item{summarizeTerms}{character vector of column expressions to summarize by.}
}
\value{
.data grouped by columns named in groupingVars
}
\description{
summarize a data frame by the summarizeTerms.  Accepts arbitrary text as
summarizeTerms to allow forms such as "mean(Sepal.Length)".
}
\examples{

suppressPackageStartupMessages(library("dplyr"))

datasets::iris \%>\%
  summarize_se(c("Mean_Sepal_Length" := "mean(Sepal.Length)",
                 "Max_Sepal_Length" := "max(Sepal.Length)")) \%>\%
  head()


}
\seealso{
\code{\link[dplyr]{summarize}}, \code{\link[dplyr]{summarize_at}}, \code{\link[seplyr]{:=}}
}
