% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmute_nse.R
\name{transmute_nse}
\alias{transmute_nse}
\title{transmute non-standard evaluation interface (deprecated).}
\usage{
transmute_nse(.data, ..., transmute_nse_env = parent.frame())
}
\arguments{
\item{.data}{data.frame}

\item{...}{stringified expressions to transmute by.}

\item{transmute_nse_env}{environment to work in.}
}
\value{
.data with altered columns(other columns dropped).
}
\description{
transmute a data frame by the transmuteterms from \code{...} (deprecated, please use \code{\link{transmute_se}}).
}
\seealso{
\code{\link{transmute_se}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr]{transmute_at}}, \code{\link[wrapr]{:=}}
}
