% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmute_se.R
\name{transmute_se}
\alias{transmute_se}
\title{transmute standard interface.}
\usage{
transmute_se(.data, transmuteTerms)
}
\arguments{
\item{.data}{data.frame}

\item{transmuteTerms}{character vector of column expressions to transmute by.}
}
\value{
.data grouped by columns named in groupingVars
}
\description{
transmute a data frame by the transmuteTerms.  Accepts arbitrary text as
transmuteTerms to allow forms such as "Sepal.Length >= 2 * Sepal.Width".
}
\examples{

suppressPackageStartupMessages(library("dplyr"))

datasets::iris \%>\%
  transmute_se(c("Sepal_Long" := "Sepal.Length >= 2 * Sepal.Width",
              "Petal_Short" := "Petal.Length <= 3.5")) \%>\%
  summary()


}
\seealso{
\code{\link[dplyr]{transmute}}, \code{\link[dplyr]{transmute_at}}, \code{\link[seplyr]{:=}}
}
