% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_se.R
\name{arrange_se}
\alias{arrange_se}
\title{arrange standard interface.}
\usage{
arrange_se(.data, arrangeTerms)
}
\arguments{
\item{.data}{data.frame}

\item{arrangeTerms}{character vector of column expressions to group by.}
}
\value{
.data grouped by columns named in groupingVars
}
\description{
Arange a data frame by the arrangeTerms.  Accepts arbitrary text as
arrangeTerms to allow forms such as "desc(gear)".
Author: John Mount, Win-Vector LLC.
}
\examples{

arrange_se(datasets::mtcars, c("cyl", "desc(gear)")) \%>\%
  head()
# roughly equivilent to:
# arrange(datasets::mtcars, cyl, desc(gear)) \%>\% head()

}
