% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringAlgebra.R
\name{expand_expr}
\alias{expand_expr}
\title{Prepare an expression for standard evaluation.}
\usage{
expand_expr(expr, env = parent.frame())
}
\arguments{
\item{expr}{to de-ref.}

\item{env}{enviornment in.}
}
\value{
string
}
\description{
Combine string-variable values and quoted terms to produce
a concrete string reading for value oriented string evaluation
(stringified standard evaluation).  We also call this the
string algebra or string interpolation.
This method powers the \code{seplyr::*_nse()}
methods and helps build string expressions without needing
\code{paste0())} or \code{glue::glue()}.
}
\examples{

resCol1 <- "Sepal_Long"
ratio <- 2
compCol1 <- "Sepal.Width"
expr <- expand_expr("Sepal.Length" >= ratio * compCol1)
print(expr)
resCol <- 'X'
iris \%.>\%
   mutate_se(., resCol1 := expr) \%.>\%
   head(.)

}
