% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_nse.R
\name{expand_expr}
\alias{expand_expr}
\title{Prepare an expression for standard evaluation.}
\usage{
expand_expr(expr, env = parent.frame())
}
\arguments{
\item{expr}{to de-ref.}

\item{env}{enviornment in.}
}
\value{
string
}
\description{
Prepare an expression for standard evaluation.
}
\examples{

resCol1 <- "Sepal_Long"
ratio <- 2
compCol1 <- "Sepal.Width"
expr <- expand_expr("Sepal.Length" >= ratio * compCol1)
print(expr)
resCol <- 'X'
head(mutate_se(iris, c(resCol := expr)))

}
