% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_nse.R
\name{filter_nse}
\alias{filter_nse}
\title{filter non-standard interface.}
\usage{
filter_nse(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{stringified expressions to filter by.}
}
\value{
.data filtered by columns named in filterTerms
}
\description{
Filter a data frame by the filterTerms.  Accepts arbitrary text as
filterTerms to allow forms such as "Sepal.Length >= 2 * Sepal.Width".
}
\examples{

upperBound <- 3.5

datasets::iris \%.>\%
  filter_nse(., "Sepal.Length" >= 2 * "Sepal.Width",
                 "Petal.Length" <= upperBound) \%.>\%
  head(.)


}
\seealso{
\code{\link{filter_se}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{filter_at}}
}
