% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_se.R
\name{mutate_se}
\alias{mutate_se}
\title{mutate standard interface.}
\usage{
mutate_se(.data, mutateTerms)
}
\arguments{
\item{.data}{data.frame}

\item{mutateTerms}{character vector of column expressions to mutate by.}
}
\value{
.data grouped by columns named in groupingVars
}
\description{
Mutate a data frame by the mutateTerms.  Accepts arbitrary text as
mutateTerms to allow forms such as "Sepal.Length >= 2 * Sepal.Width".
Terms are vectors or lists of the form
}
\examples{

datasets::iris \%>\%
  mutate_se(c("Sepal_Long" := "Sepal.Length >= 2 * Sepal.Width",
              "Petal_Short" := "Petal.Length <= 3.5")) \%>\%
  summary()


}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate_at}}, \code{\link[seplyr]{:=}}
}
