% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_se.R
\name{mutate_se}
\alias{mutate_se}
\title{mutate standard evaluation interface.}
\usage{
mutate_se(.data, mutateTerms, ..., splitTerms = TRUE, warn = TRUE,
  env = parent.frame(), printPlan = FALSE)
}
\arguments{
\item{.data}{data.frame}

\item{mutateTerms}{character vector of column expressions to mutate by.}

\item{...}{force later terms to be bound by name}

\item{splitTerms}{logical, if TRUE into separate mutates (if FALSE instead, pass all at once to dplyr).}

\item{warn}{logical, if TRUE warn about name re-use.}

\item{env}{environment to work in.}

\item{printPlan}{logical, if TRUE print the expression plan.}
}
\value{
.data with altered columns.
}
\description{
Mutate a data frame by the mutateTerms.  Accepts arbitrary text as
mutateTerms to allow forms such as "Sepal.Length >= 2 * Sepal.Width".
Terms are vectors or lists of the form "lhs := rhs".
Semantics are: terms are evaluated left to right if splitTerms==TRUE (the default).
}
\details{
Note: this method as the default setting \code{splitTerms = TRUE}, which is
safer (avoiding certain known \code{dplyr}/\code{dblyr} issues)
(please see the side-notes of \url{http://winvector.github.io/FluidData/partition_mutate.html} for some references).
}
\examples{


limit <- 3.5

datasets::iris \%.>\%
  mutate_se(., qae(Sepal_Long = Sepal.Length >= 2 * Sepal.Width,
                   Petal_Short := Petal.Length <= limit)) \%.>\%
  head(.)


}
\seealso{
\code{\link{mutate_nse}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate_at}}, \code{\link[wrapr]{:=}}
}
