% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringAlgebra.R
\name{prep_deref}
\alias{prep_deref}
\title{Substitute language elements by one-expand_expr (deprecated).}
\usage{
prep_deref(lexpr, env = parent.frame())
}
\arguments{
\item{lexpr}{language item (captured by substitute())}

\item{env}{environment to look in}
}
\value{
R language element with dequotes and derefs
}
\description{
Substitute language elements by one-expand_expr (deprecated).
}
