% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_se.R
\name{rename_se}
\alias{rename_se}
\title{rename standard interface.}
\usage{
rename_se(.data, mapping)
}
\arguments{
\item{.data}{data.frame}

\item{mapping}{named character vector of columns to rename (new news on the left, original names on the right).}
}
\value{
.data with renamed columns
}
\description{
rename columns.
}
\examples{

datasets::mtcars \%>\% rename_se(c("cylinders" = "cyl", "gears" = "gear")) \%>\% head()

}
