% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_se.R
\name{select_se}
\alias{select_se}
\title{Select columns standard interface.}
\usage{
select_se(.data, colNames)
}
\arguments{
\item{.data}{data.frame}

\item{colNames}{character vector of columns to keep}
}
\value{
.data with only selected columns
}
\description{
Select columns.  To remove columns please see \code{\link{deselect}}.  Also accepts -column notation.
}
\examples{


datasets::mtcars \%.>\%
   select_se(., c("cyl", "gear")) \%.>\%
   head(.)
# essentially dplyr::select_at()

data.frame(a=1, b=2) \%.>\% select_se(., '-b')

}
\seealso{
\code{\link{deselect}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{select_at}}
}
