% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{seplyr}
\alias{seplyr}
\alias{seplyr-package}
\title{\code{seplyr}: Standard Evaluation Improved Interfaces for Common Data Manipulatio Tasks}
\description{
The \code{seplyr} (standard evaluation \code{dplyr}) package supplies improved
standard evaluation adapter methods for important common data manipulation tasks.
}
\details{
In addition the \code{seplyr} package supplies several new "key operations
bound together" methods.  These include \code{group_summarize()} (which
combines grouping, arranging and calculation in an atomic unit),
\code{add_group_summaries()} (which joins grouped summaries into a \code{data.frame}
in a well documented manner), \code{add_group_indices()} (which adds
per-group identifiers to a \code{data.frame} without depending on row-order),
\code{partition_mutate_qt()} (which optimizes mutate sequences), and \code{if_else_device()}
(which simulates per-row if-else blocks in expression sequences).
}
