% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_nse.R
\name{summarize_nse}
\alias{summarize_nse}
\alias{summarise_nse}
\title{summarize non-standard evaluation interface (deprecated).}
\usage{
summarize_nse(.data, ..., env = parent.frame())

summarise_nse(.data, ..., env = parent.frame())
}
\arguments{
\item{.data}{data.frame}

\item{...}{stringified expressions to summarize by.}

\item{env}{environment to work in.}
}
\value{
.data with summarized columns.
}
\description{
summarize a data frame by the summarize terms from \code{...} (deprecated, please use \code{\link{summarize_se}}).
}
\seealso{
\code{\link{summarize_se}}, \code{\link[dplyr]{summarize}}, \code{\link[dplyr]{summarize_at}}, \code{\link[wrapr]{:=}}
}
