% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_nse.R
\name{summarize_nse}
\alias{summarize_nse}
\alias{summarise_nse}
\title{summarize non-standard evaluation interface.}
\usage{
summarize_nse(.data, ..., summarize_nse_warn = TRUE,
  env = parent.frame())

summarise_nse(.data, ..., summarize_nse_warn = TRUE,
  env = parent.frame())
}
\arguments{
\item{.data}{data.frame}

\item{...}{stringified expressions to summarize by.}

\item{summarize_nse_warn}{logical, if TRUE warn about possible name collisions.}

\item{env}{environment to work in.}
}
\value{
.data with summarized columns.
}
\description{
summarize a data frame by the summarize terms from \code{...}.
}
\examples{


datasets::iris \%.>\%
  summarize_nse(., Mean_Sepal_Length := mean(Sepal.Length),
                   Max_Sepal_Length := max(Sepal.Length))

}
\seealso{
\code{\link{summarize_se}}, \code{\link[dplyr]{summarize}}, \code{\link[dplyr]{summarize_at}}, \code{\link[wrapr]{:=}}
}
